+++ main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"fmt"
 	"inv/cmd/dig"
-	"inv/cmd/up"
+	"inv/cmd/ping"
 	"inv/cmd/whois"
 	"os"
 
@@ -13,7 +13,7 @@ import (
 
 var (
 	digVar   string
-	upVar    string
+	pingVar  string
 	whoisVar string
 	version  string
 	banner   = "Usage: inv [OPTIONS] COMMAND\n"
@@ -22,8 +22,8 @@ var (
 func main() {
 	op := optionparser.NewOptionParser()
 	op.Banner = banner
-	op.On("-u", "--up IP", "Check liveness stats", &upVar)
 	op.On("-d", "--dig domain", "Dig DNS records", &digVar)
+	op.On("-p", "--ping IP", "Check liveness stats", &pingVar)
 	op.On("-w", "--whois domain", "Whois domain information", &whoisVar)
 	op.On("--version", "Print version")
 
@@ -42,8 +42,8 @@ func main() {
 		dig.Dig(digVar)
 	case "-w", "-whois", "--whois":
 		whois.Whois(whoisVar)
-	case "-u", "-up", "--up":
-		up.Up(upVar)
+	case "-p", "-ping", "--ping":
+		ping.Ping(pingVar)
 	case "--version":
 		fmt.Printf("Version: %s\n", version)
 	default:
